leaf-disease-webapp/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pillow==10.4.0
torch==2.3.1
numpy<2
